pyproject.toml
[tool.poetry]
name = "shortenerlinks"
version = "0.1.0"
description = ""
authors = ["N1chegons <krikignatenko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.15"
fastapi = "^0.125.0"
uvicorn = "^0.38.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^15.0.3"}
asyncpg = "^0.31.0"
sqlalchemy = "^2.0.45"
alembic = "^1.17.2"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.2"}
pydantic-settings = "^2.12.0"
pytest = "^9.0.2"
httpx = "^0.28.1"
aiosqlite = "^0.22.0"
pytest-asyncio = "^1.3.0"
segno = "^1.6.6"
resend = "^2.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
